% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-bbox.R
\name{geo_bbox}
\alias{geo_bbox}
\alias{geo_bbox.default}
\alias{geo_bbox.geo_collection}
\title{Bounding boxes}
\usage{
geo_bbox(x, ..., na.rm = FALSE)

\method{geo_bbox}{default}(x, ..., na.rm = FALSE)

\method{geo_bbox}{geo_collection}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A geometry-like object}

\item{...}{Unused}

\item{na.rm}{Should NAs be considered?}
}
\value{
A \code{\link[=geo_rect]{geo_rect()}} of length 1.
}
\description{
Bounding boxes
}
\examples{
geo_bbox(geo_wkt(c("POINT (30 10)", "POINT EMPTY")))

}
