% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coalesce.R
\name{geo_coalesce}
\alias{geo_coalesce}
\title{Find first non-empty feature}
\usage{
geo_coalesce(x, ...)
}
\arguments{
\item{x, ...}{Vectors that will be recycled to a common length. These don't
necessarily need to be geovctrs, but do need methods for \code{\link[=geo_is_empty]{geo_is_empty()}}
and must be castable to the type of the first argument.}
}
\value{
A vector with the type of the first argument.
}
\description{
Like \code{\link[dplyr:coalesce]{dplyr::coalesce()}}, \code{\link[=geo_coalesce]{geo_coalesce()}} returns the first non-missing
element (rowwise). It is useful for replacing empty and missing values
with a default.
}
\examples{
wkt <- wkt(c("POINT (30 10)", "POINT EMPTY", NA))
geo_coalesce(wkt, wkt("POINT (0 0)"))

}
