% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coordinates.R
\name{geo_coordinates}
\alias{geo_coordinates}
\title{Extract coordinates as a tibble}
\usage{
geo_coordinates(x, ...)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Unused}
}
\value{
A tibble with one row per coordinate and columns
\code{feature} and \code{xy}. Vectors that include mutli geometries
will have a \code{part} column, and vectors that include polygons
will have a \code{ring} column.
}
\description{
Extract coordinates as a tibble
}
\examples{
geo_coordinates("POINT (20 17)")
geo_coordinates("POINT EMPTY")

geo_coordinates("LINESTRING EMPTY")
geo_coordinates("LINESTRING (30 10, 0 0)")

geo_coordinates("POLYGON EMPTY")
geo_coordinates("POLYGON ((30 10, 0 0, -30 10, 30 10))")

geo_coordinates(
  "
  MULTIPOLYGON (
    ((40 40, 20 45, 45 30, 40 40)),
    ((20 35, 10 30, 10 10, 30 5, 45 20, 20 35),
      (30 20, 20 15, 20 25, 30 20))
  )"
)

}
