% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_example_wkt}
\alias{geo_example_wkt}
\title{Example well-known text}
\format{An object of class \code{geo_wkt} (inherits from \code{geovctr}, \code{vctrs_vctr}) of length 19.}
\source{
\href{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}{Wikipedia's well-known text page}
}
\usage{
geo_example_wkt
}
\description{
Various examples of well-known text useful for testing.
}
\keyword{datasets}
