% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-geometry-type.R
\name{geo_geometry_type}
\alias{geo_geometry_type}
\alias{geo_geometry_type.default}
\alias{geo_geometry_type.wk_wkt}
\alias{geo_geometry_type.wk_wkb}
\alias{geo_geometry_type.wk_wksxp}
\title{Extract feature geometry type}
\usage{
geo_geometry_type(x)

\method{geo_geometry_type}{default}(x)

\method{geo_geometry_type}{wk_wkt}(x)

\method{geo_geometry_type}{wk_wkb}(x)

\method{geo_geometry_type}{wk_wksxp}(x)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}
}
\value{
A character vector with lowercase geometry types
}
\description{
Extract feature geometry type
}
\examples{
geo_geometry_type("POINT EMPTY")

}
