% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-grob.R
\name{geo_grob}
\alias{geo_grob}
\alias{geo_grob.default}
\alias{geo_grob.wk_wkt}
\alias{geo_grob.wk_wkb}
\alias{geo_grob.wk_wksxp}
\alias{geo_grob.geovctrs_xy}
\alias{geo_grob.geovctrs_segment}
\alias{geo_grob.geovctrs_rect}
\title{Create graphical objects}
\usage{
geo_grob(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{default}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{wk_wkt}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{wk_wkb}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{wk_wksxp}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{geovctrs_xy}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{geovctrs_segment}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)

\method{geo_grob}{geovctrs_rect}(
  x,
  ...,
  rule = "evenodd",
  default.units = "native",
  name = NULL,
  vp = NULL
)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Graphical parameters passed to \code{\link[grid:gpar]{grid::gpar()}}. These are recycled along
the input. Dynamic dots (e.g., \verb{!!!}) are supported.}

\item{rule}{Use "winding" if polygon rings are correctly encoded with a winding
direction.}

\item{default.units}{Coordinate units, which may be defined by the viewport (see
\code{\link[grid:unit]{grid::unit()}}). Defaults to native.}

\item{name}{Passed to \code{\link[grid:grid.points]{grid::pointsGrob()}}, \code{\link[grid:grid.lines]{grid::polylineGrob()}},
\code{\link[grid:grid.path]{grid::pathGrob()}}, or \code{\link[grid:grid-defunct]{grid::gTree()}} depending on the types of geometries
in the input.}

\item{vp}{Passed to \code{\link[grid:grid.points]{grid::pointsGrob()}}, \code{\link[grid:grid.lines]{grid::polylineGrob()}},
\code{\link[grid:grid.path]{grid::pathGrob()}}, or \code{\link[grid:grid-defunct]{grid::gTree()}} depending on the types of geometries
in the input.}
}
\value{
A \link[grid:grid.grob]{grid graphical object}
}
\description{
Create graphical objects
}
\examples{
grid::grid.newpage()
grid::grid.draw(
  geo_grob("POINT (0.5 0.5)", col = "red", pch = 16, default.units = "npc")
)

}
