% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-missing.R
\name{geo_is_missing}
\alias{geo_is_missing}
\alias{geo_has_missing}
\alias{geo_is_finite}
\title{Missing and non-finite geometries}
\usage{
geo_is_missing(x)

geo_has_missing(x)

geo_is_finite(x)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}
}
\value{
A logical vector
}
\description{
Missing and non-finite geometries
}
