% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-lim.R
\name{geo_lim}
\alias{geo_lim}
\title{One-dimensional limits}
\usage{
geo_lim(lower = double(), upper = double())
}
\arguments{
\item{lower, upper}{The lower and upper bounds of the one-dimensional rage.}
}
\value{
A \code{\link[=new_geovctrs_lim]{new_geovctrs_lim()}}
}
\description{
Vectorized version of \code{range()} to support \code{\link[=geo_xlim]{geo_xlim()}},
\code{\link[=geo_ylim]{geo_ylim()}}, and \code{\link[=geo_zlim]{geo_zlim()}}.
}
\examples{
geo_lim(3, 4)

}
