% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-linestring.R
\name{geo_linestring}
\alias{geo_linestring}
\title{Create a linestring}
\usage{
geo_linestring(xy, srid = 0)
}
\arguments{
\item{xy}{A \code{\link[=geo_xy]{geo_xy()}}}

\item{srid}{A spatial reference identifier, coerced to
an integer by \code{\link[=as_geo_srid]{as_geo_srid()}}. These identifiers can
and should be managed outside of geovctrs except for
0, which is interpreted as "not set".}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a linestring
}
\examples{
geo_linestring(geo_xy(1:5, 2:6))

}
