% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-multilinestring.R
\name{geo_multilinestring}
\alias{geo_multilinestring}
\title{Create a multilinestring}
\usage{
geo_multilinestring(feature, srid = NA)
}
\arguments{
\item{feature}{A \code{\link[=geo_collection]{geo_collection()}} of \code{\link[=geo_linestring]{geo_linestring()}}s.}

\item{srid}{A spatial reference identifier.}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a multilinestring
}
\examples{
geo_multilinestring(
  geo_linestring(geo_xy(0:1, 0:1))
)

}
