% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-multipolygon.R
\name{geo_multipolygon}
\alias{geo_multipolygon}
\title{Create a multipolygon}
\usage{
geo_multipolygon(feature, srid = 0)
}
\arguments{
\item{feature}{A \code{\link[=geo_collection]{geo_collection()}} of \code{\link[=geo_polygon]{geo_polygon()}}s.}

\item{srid}{A spatial reference identifier, coerced to
an integer by \code{\link[=as_geo_srid]{as_geo_srid()}}. These identifiers can
and should be managed outside of geovctrs except for
0, which is interpreted as "not set".}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a multipolygon
}
\examples{
geo_multipolygon(
  geo_polygon(geo_xy(c(0, 10, 0, 0), c(0, 0, 10, 0)))
)

}
