% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_nc}
\alias{geo_nc}
\title{North Carolina dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 12 columns.
}
\source{
The \link[sf:sf]{sf package}.
}
\usage{
geo_nc
}
\description{
A spatial data package wouldn't be complete without a copy of the NC
dataset.
}
\keyword{datasets}
