% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-polygon.R
\name{geo_polygon}
\alias{geo_polygon}
\title{Create a polygon}
\usage{
geo_polygon(xy, ring = 1L, srid = NA)
}
\arguments{
\item{xy}{A \code{\link[=geo_xy]{geo_xy()}}}

\item{srid}{A spatial reference identifier.}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a polygon
}
\examples{
geo_polygon(geo_xy(c(0, 10, 0), c(0, 0, 10)))

}
