% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-print.R
\name{geo_print}
\alias{geo_print}
\alias{geo_format}
\title{Format and print geovctrs}
\usage{
geo_print(x, ..., col = TRUE)

geo_format(x, ..., col = FALSE)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Passed to \code{\link[=geo_format]{geo_format()}}}

\item{col}{Use to colourize the output}

\item{short}{Use to print a shorter summary}
}
\description{
Format and prints a geovctr.
}
\examples{
geo_format(head(geo_nc))
geo_print(head(geo_nc))

}
