% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-segment.R
\name{geo_segment}
\alias{geo_segment}
\title{Line segments}
\usage{
geo_segment(
  x0 = double(),
  y0 = double(),
  x1 = double(),
  y1 = double(),
  srid = NA
)
}
\arguments{
\item{x0, y0, x1, y1}{Values for the start and end coordinates.}

\item{srid}{A spatial reference identifier, coerced to
an integer by \code{\link[=as_geo_srid]{as_geo_srid()}}. These identifiers can
and should be managed outside of geovctrs except for
0, which is interpreted as "not set".}
}
\value{
A \code{\link[=new_geovctrs_segment]{new_geovctrs_segment()}}
}
\description{
The \code{\link[=geo_segment]{geo_segment()}} type is useful as an efficient representation of
line segments stored using column vectors.
}
\examples{
plot(geo_segment(0, 0, 10, -10:10))

}
