% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-z.R
\name{geo_set_z}
\alias{geo_set_z}
\alias{geo_drop_z}
\alias{geo_has_z}
\title{Work with Z values}
\usage{
geo_set_z(x, z)

geo_drop_z(x)

geo_has_z(x)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{z}{A z value}
}
\value{
A modified version of \code{x}.
}
\description{
Currently, geovctrs supports X, Y, and Z coordinates. These functions help extract,
set, and drop Z values.
}
\examples{
geo_set_z("POINT (2 3)", 4)
geo_drop_z("POINT Z (2 3 4)")

}
