% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-srid.R
\name{geo_srid}
\alias{geo_srid}
\alias{set_geo_srid}
\alias{as_geo_srid}
\title{Extract a spatial reference identifier}
\usage{
geo_srid(x)

set_geo_srid(x, srid)

as_geo_srid(x)
}
\arguments{
\item{x}{A geometry-like object.}

\item{srid}{A spatial reference identifier, coerced to
an integer by \code{\link[=as_geo_srid]{as_geo_srid()}}. These identifiers can
and should be managed outside of geovctrs.}
}
\value{
An integer vector (one SRID per feature).
}
\description{
The geovctrs package works with spatial reference identifiers
instead of actual CRS objects so that the heavy lifting can
be implemented in other packages. However, most geometry
formats store an integer spatial reference identifier (SRID)
with each feature to propogate this information through
calculations. Note that an SRID of 0 is ambiguous and is
not allowed (use \code{NA} to represent an undefined SRID).
}
\examples{
# two points with an SRID
geometries <- c(
  geo_point(geo_xy(259473, 4876249), srid = 26920),
  geo_point(geo_xy(-66, 44), srid = 4326)
)

geo_srid(geometries)
geo_srid(set_geo_srid(geometries, NA))

# SRIDs are propogated through conversions,
# or discarded with a warning
geo_srid(as_geo_wkb(geometries))
geo_srid(as_geo_xy(geometries))

}
