% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-summary.R
\name{geo_summary}
\alias{geo_summary}
\alias{geo_n_geometries}
\alias{geo_n_coordinates}
\alias{geo_geometry_type}
\alias{geo_coordinate_dimensions}
\alias{geo_first_coordinate}
\title{Feature-level geometry summaries}
\usage{
geo_summary(x)

geo_n_geometries(x)

geo_n_coordinates(x)

geo_geometry_type(x)

geo_coordinate_dimensions(x)

geo_first_coordinate(x)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{geometry_type}, \code{is_empty},
\code{n_coordinates}, \code{srid}, \code{coordinate_dimensions}, and \code{envelope}.
}
\description{
Feature-level geometry summaries
}
\examples{
geo_summary(
  geo_wkt(
    c(
    "POINT (30 10)",
    "LINESTRING (0 0, 10 11)",
    "POLYGON EMPTY"
   )
 )
)

}
