% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-transform.R
\name{geo_transform}
\alias{geo_transform}
\alias{geo_transform.default}
\alias{geo_transform.wk_wkt}
\alias{geo_transform.wk_wkb}
\alias{geo_transform.wk_wksxp}
\alias{geo_transform.geovctrs_xy}
\alias{geo_transform.geovctrs_segment}
\title{Apply an affine transformation}
\usage{
geo_transform(x, trans, ...)

\method{geo_transform}{default}(x, trans, ...)

\method{geo_transform}{wk_wkt}(x, trans, ...)

\method{geo_transform}{wk_wkb}(x, trans, ...)

\method{geo_transform}{wk_wksxp}(x, trans, ...)

\method{geo_transform}{geovctrs_xy}(x, trans, ...)

\method{geo_transform}{geovctrs_segment}(x, trans, ...)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{trans}{A 3x3 transformation matrix describing an affine
transformation of the input.}

\item{...}{Unused}
}
\value{
A transformed \code{x}
}
\description{
Apply an affine transformation
}
\examples{
geo_transform(
  "POINT (30 10)",
  # translation +12 +13
  matrix(c(1, 0, 0, 0, 1, 0, 12, 13, 1), ncol = 3)
)

geo_transform(
  geo_nc,
  # translation +12 +13
  matrix(c(1, 0, 0, 0, 1, 0, 12, 13, 1), ncol = 3)
)[c("NAME", "geometry")]

}
