% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-unnest.R
\name{geo_unnest}
\alias{geo_unnest}
\alias{geo_unnest.default}
\alias{geo_unnest.wk_wkt}
\alias{geo_unnest.wk_wkb}
\alias{geo_unnest.wk_wksxp}
\title{Unnest recursive geometry structures}
\usage{
geo_unnest(x, ..., keep_empty = FALSE, keep_multi = TRUE, max_depth = 1)

\method{geo_unnest}{default}(x, ..., keep_empty = FALSE, keep_multi = TRUE, max_depth = 1)

\method{geo_unnest}{wk_wkt}(x, ..., keep_empty = FALSE, keep_multi = TRUE, max_depth = 1)

\method{geo_unnest}{wk_wkb}(x, ..., keep_empty = FALSE, keep_multi = TRUE, max_depth = 1)

\method{geo_unnest}{wk_wksxp}(x, ..., keep_empty = FALSE, keep_multi = TRUE, max_depth = 1)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Unused}

\item{keep_empty}{If \code{TRUE}, a GEOMETRYCOLLECTION EMPTY is left as-is
rather than collapsing to length 0.}

\item{keep_multi}{If \code{TRUE}, MULTI* geometries are not expanded to sub-features.}

\item{max_depth}{The maximum recursive GEOMETRYCOLLECTION depth to unnest.}
}
\description{
Unnest recursive geometry structures
}
\examples{
geo_unnest("GEOMETRYCOLLECTION (POINT (1 2), POINT (3 4))")
geo_unnest("GEOMETRYCOLLECTION EMPTY")
geo_unnest("GEOMETRYCOLLECTION EMPTY", keep_empty = TRUE)

geo_unnest(geo_example_wkt, keep_multi = TRUE)
geo_unnest(geo_example_wkt, keep_multi = FALSE)

}
