% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-unnest.R
\name{geo_unnest_collection}
\alias{geo_unnest_collection}
\alias{geo_unnest_all}
\alias{geo_unnest_all.default}
\alias{geo_unnest_all.wk_wkt}
\alias{geo_unnest_all.wk_wkb}
\alias{geo_unnest_all.wk_wksxp}
\title{Unnest recursive geometry structures}
\usage{
geo_unnest_collection(x, keep_empty = FALSE)

geo_unnest_all(x, keep_empty = FALSE)

\method{geo_unnest_all}{default}(x, keep_empty = FALSE)

\method{geo_unnest_all}{wk_wkt}(x, keep_empty = FALSE)

\method{geo_unnest_all}{wk_wkb}(x, keep_empty = FALSE)

\method{geo_unnest_all}{wk_wksxp}(x, keep_empty = FALSE)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{keep_empty}{If \code{TRUE}, a GEOMETRYCOLLECTION EMPTY is left as-is
rather than collapsing to length 0}
}
\description{
Unnest recursive geometry structures
}
\examples{
geo_unnest_collection("GEOMETRYCOLLECTION (POINT (1 2), POINT (3 4))")
geo_unnest_collection("GEOMETRYCOLLECTION EMPTY")
geo_unnest_collection("GEOMETRYCOLLECTION EMPTY", keep_empty = TRUE)

}
