% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkb.R
\name{geo_wkb}
\alias{geo_wkb}
\alias{parse_wkb}
\alias{as_geo_wkb}
\title{Create and validate well-known binary}
\usage{
geo_wkb(x = list())

parse_wkb(x)

as_geo_wkb(x, ..., include_srid = NA, dimensions = 3, endian = NA)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} objects, each of which
represent well-known binary}

\item{...}{Unused}

\item{include_srid}{Use \code{TRUE} to always include, \code{FALSE}
to never include, or \code{NA} to only include if the SRID
is non-zero.}

\item{dimensions}{The maximum number of dimensions to consider
when generating text.}

\item{endian}{Use \code{0} for big endian, \code{1} for little endian
or \code{NA} to use the default on your system.}
}
\value{
A \code{\link[=new_geovctrs_wkb]{new_geovctrs_wkb()}}
}
\description{
Like other geo types, \code{\link[=geo_wkb]{geo_wkb()}} doesn't convert its input
but does validate it using \code{\link[=validate_geovctrs_wkb]{validate_geovctrs_wkb()}}.
To skip validation, use \code{\link[=new_geovctrs_wkb]{new_geovctrs_wkb()}} with
the result of \code{vec_cast(list(...), .ptype = raw())}.
}
\examples{
# POINT (30 10) in WKB
wkb_item <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
wkb <- geo_wkb(list(wkb_item))
wkb

# use as_geo_wkb() to use conversion options
as_geo_wkb(wkb, endian = 0)

# use parse_wkb() to identify parse errors
wkb_item[2] <- as.raw(0x11)
parse_wkb(list(wkb_item))

}
