% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-xy.R
\name{geo_xy}
\alias{geo_xy}
\title{Create a coordinate vector}
\usage{
geo_xy(x = double(), y = double())
}
\arguments{
\item{x, y}{x and y coordinates}
}
\value{
A \code{\link[=new_geo_xy]{new_geo_xy()}}
}
\description{
The \code{\link[=geo_xy]{geo_xy()}} type is useful as an efficient representation of points
stored using column vectors. Note that \code{geo_xy(NA, NA)} is considered
an "empty" point rather than a "missing" point (see \code{\link[=geo_is_missing]{geo_is_missing()}}
and \code{\link[=geo_is_empty]{geo_is_empty()}}).
}
\examples{
geo_plot(geo_xy(0:5, 1:6))

}
