% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-version.R
\name{geos_version_runtime}
\alias{geos_version_runtime}
\alias{geos_version_build}
\title{GEOS version information}
\usage{
geos_version_runtime()

geos_version_build()
}
\description{
GEOS version information
}
\examples{
geos_version_runtime()
geos_version_build()

}
