% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geovctrs-package.R
\docType{package}
\name{geovctrs-package}
\alias{geovctrs}
\alias{geovctrs-package}
\title{geovctrs: Efficient Data Structures for Geometry Vectors}
\description{
Provides a set of data structures
  to efficiently encode simple geometries such as points,
  segments, and rectangles. Coercion functions are provided
  such that any vector of geometry-like objects can be used
  interchangeably.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/geovctrs}
  \item \url{https://github.com/paleolimbot/geovctrs}
  \item Report bugs at \url{https://github.com/paleolimbot/geovctrs/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
