% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geovctr.R
\name{is_geovctr}
\alias{is_geovctr}
\alias{as_geovctr}
\alias{as_geovctr.geovctr}
\alias{as_geovctr.character}
\alias{expect_geovctr}
\title{Test for geovctrs}
\usage{
is_geovctr(x)

as_geovctr(x, ...)

\method{as_geovctr}{geovctr}(x, ...)

\method{as_geovctr}{character}(x, ...)

expect_geovctr(x)
}
\arguments{
\item{x}{A (possibly) geovctr}

\item{...}{Passed to the constructor}
}
\description{
A geovctr is a geometry-like object that can be converted to
the standard for geometry, well-known binary. In the geovctrs
package, \code{\link[=geo_wkt]{geo_wkt()}}, \code{\link[=geo_wkb]{geo_wkb()}}, \code{\link[=geo_collection]{geo_collection()}},
\code{\link[=geo_xy]{geo_xy()}}, \code{\link[=geo_rect]{geo_rect()}}, and \code{\link[=geo_segment]{geo_segment()}} are all geovctrs.
Extension packages can implement the \code{\link[=as_geovctr]{as_geovctr()}} generic
to take advantage of the functions in this package, including
coercion, plotting, and summary information. An example of this
is \code{as_geovctr.character()}, which allow \code{\link[=geo_plot]{geo_plot()}} to be called
on a character vector of well-known text.
}
\examples{
is_geovctr(geo_wkt())
is_geovctr(NULL)

as_geovctr(geo_wkt("POINT (30 10)"))
as_geovctr("POINT (30 10)")

}
