% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkb.R
\name{is_wk_wkb}
\alias{is_wk_wkb}
\alias{validate_wk_wkb}
\alias{vec_cast.wk_wkb}
\alias{vec_ptype2.wk_wkb}
\title{S3 details for wk_wkb}
\usage{
is_wk_wkb(x)

validate_wk_wkb(x)

\method{vec_cast}{wk_wkb}(x, to, ...)

\method{vec_ptype2}{wk_wkb}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} objects, each of which
represent well-known binary}

\item{to}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{...}{Unused}

\item{y}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for wk_wkb
}
\examples{
wkb_raw <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
wkb <- geo_wkb(list(wkb_raw))
is_wk_wkb(wkb)

# to skip parse validation if you know your object is
# valid, use new_wk_wkb()
new_wk_wkb(list(wkb_raw))

}
