% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkt.R
\name{is_wk_wkt}
\alias{is_wk_wkt}
\alias{vec_cast.wk_wkt}
\alias{vec_ptype2.wk_wkt}
\title{S3 details for wk_wkt}
\usage{
is_wk_wkt(x)

\method{vec_cast}{wk_wkt}(x, to, ...)

\method{vec_ptype2}{wk_wkt}(x, y, ...)
}
\arguments{
\item{x}{A character vector containing well-known text}

\item{...}{Unused}

\item{y, to}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for wk_wkt
}
\examples{
wkt <- geo_wkt("POINT (30 10)")
is_wk_wkt(wkt)

}
