% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coord-collection.R
\name{new_geo_coord_collection}
\alias{new_geo_coord_collection}
\alias{validate_geo_coord_collection}
\alias{vec_ptype_abbr.geo_coord_collection}
\alias{format.geo_coord_collection}
\alias{print.geo_coord_collection}
\alias{is_geo_coord_collection}
\title{S3 Details for coordinate vector collections}
\usage{
new_geo_coord_collection(x, srid = NA_integer_)

validate_geo_coord_collection(x)

\method{vec_ptype_abbr}{geo_coord_collection}(x, ...)

\method{format}{geo_coord_collection}(x, ...)

\method{print}{geo_coord_collection}(x, ...)

is_geo_coord_collection(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=list]{list()}} of geo coordinate vectors}

\item{srid}{A spatial reference identifier}

\item{...}{Unused}
}
\description{
S3 Details for coordinate vector collections
}
