% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkb.R
\name{new_geo_wkb}
\alias{new_geo_wkb}
\alias{is_geo_wkb}
\alias{validate_geo_wkb}
\alias{as_geo_wkb}
\alias{as_geo_wkb.default}
\alias{vec_cast.geo_wkb}
\alias{vec_ptype2.geo_wkb}
\title{S3 details for geo_wkb}
\usage{
new_geo_wkb(x = vctrs::list_of(.ptype = raw()))

is_geo_wkb(x)

validate_geo_wkb(x)

as_geo_wkb(x, ...)

\method{as_geo_wkb}{default}(x, ...)

\method{vec_cast}{geo_wkb}(x, to, ...)

\method{vec_ptype2}{geo_wkb}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} objects, each of which
represent well-known binary}

\item{...}{Unused}

\item{to}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{y}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geo_wkb
}
\examples{
wkb_raw <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
wkb <- geo_wkb(list(wkb_raw))
is_geo_wkb(wkb)

}
