% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-xy.R
\name{new_geo_xy}
\alias{new_geo_xy}
\alias{is_geo_xy}
\alias{as_geo_xy.matrix}
\alias{as.matrix.geo_xy}
\alias{as_geo_xy}
\alias{as_geo_xy.default}
\alias{vec_cast.geo_xy}
\alias{vec_ptype2.geo_xy}
\title{S3 details for geo_xy}
\usage{
new_geo_xy(x = list(x = double(), y = double()))

is_geo_xy(x)

\method{as_geo_xy}{matrix}(x, ...)

\method{as.matrix}{geo_xy}(x, ...)

as_geo_xy(x, ...)

\method{as_geo_xy}{default}(x, ...)

\method{vec_cast}{geo_xy}(x, to, ...)

\method{vec_ptype2}{geo_xy}(x, y, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_xy]{geo_xy()}}}

\item{...}{Unused}

\item{to, y}{Arguments to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geo_xy
}
