% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-lim.R
\name{new_geovctrs_lim}
\alias{new_geovctrs_lim}
\alias{is_geovctrs_lim}
\alias{as_geo_lim}
\alias{as_geo_lim.matrix}
\alias{as.matrix.geovctrs_lim}
\title{S3 details for geovctrs_lim}
\usage{
new_geovctrs_lim(x = list(lower = double(), upper = double()))

is_geovctrs_lim(x)

as_geo_lim(x, ...)

\method{as_geo_lim}{matrix}(x, ...)

\method{as.matrix}{geovctrs_lim}(x, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_lim]{geo_lim()}}}

\item{...}{Unused}
}
\description{
S3 details for geovctrs_lim
}
