% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-rect.R
\name{new_geovctrs_rect}
\alias{new_geovctrs_rect}
\alias{is_geovctrs_rect}
\alias{as_geo_rect}
\alias{as_geo_rect.default}
\alias{vec_cast.geovctrs_rect}
\alias{vec_ptype2.geovctrs_rect}
\title{S3 details for geovctrs_rect}
\usage{
new_geovctrs_rect(
  x = list(xmin = double(), ymin = double(), xmax = double(), ymax = double(), srid =
    integer())
)

is_geovctrs_rect(x)

as_geo_rect(x, ...)

\method{as_geo_rect}{default}(x, ...)

\method{vec_cast}{geovctrs_rect}(x, to, ...)

\method{vec_ptype2}{geovctrs_rect}(x, y, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_rect]{geo_rect()}}}

\item{...}{Unused}

\item{to}{Arguments to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{y}{Arguments to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geovctrs_rect
}
