% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-segment.R
\name{new_geovctrs_segment}
\alias{new_geovctrs_segment}
\alias{is_geovctrs_segment}
\alias{validate_geovctrs_segment}
\alias{as_geo_segment}
\alias{as_geo_segment.default}
\alias{vec_cast.geovctrs_segment}
\alias{vec_ptype2.geovctrs_segment}
\title{S3 details for geovctrs_segment}
\usage{
new_geovctrs_segment(
  x = list(x0 = double(), y0 = double(), x1 = double(), y1 = double(), srid =
    integer())
)

is_geovctrs_segment(x)

validate_geovctrs_segment(x)

as_geo_segment(x, ...)

\method{as_geo_segment}{default}(x, ...)

\method{vec_cast}{geovctrs_segment}(x, to, ...)

\method{vec_ptype2}{geovctrs_segment}(x, y, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_segment]{geo_segment()}}}

\item{...}{Unused}

\item{to}{Arguments to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{y}{Arguments to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geovctrs_segment
}
