% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkb.R
\name{new_geovctrs_wkb}
\alias{new_geovctrs_wkb}
\alias{is_geo_wkb}
\alias{validate_geo_wkb}
\alias{vec_cast.geovctrs_wkb}
\alias{vec_ptype2.geovctrs_wkb}
\title{S3 details for geo_wkb}
\usage{
new_geovctrs_wkb(
  x = list(),
  include_srid = NA,
  dimensions = 3L,
  endian = NA_integer_
)

is_geo_wkb(x)

validate_geo_wkb(x)

\method{vec_cast}{geovctrs_wkb}(x, to, ...)

\method{vec_ptype2}{geovctrs_wkb}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} objects, each of which
represent well-known binary}

\item{include_srid}{Use \code{TRUE} to always include, \code{FALSE}
to never include, or \code{NA} to only include if the SRID
is non-zero.}

\item{dimensions}{The maximum number of dimensions to consider
when generating text.}

\item{endian}{Use \code{0} for big endian, \code{1} for little endian
or \code{NA} to use the default on your system.}

\item{to}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{...}{Unused}

\item{y}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geo_wkb
}
\examples{
wkb_raw <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
wkb <- geo_wkb(list(wkb_raw))
is_geo_wkb(wkb)

# to skip parse validation if you know your object is
# valid, use new_geovctrs_wkb()
new_geovctrs_wkb(list(wkb_raw))

}
