% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-wkt.R
\name{new_geovctrs_wkt}
\alias{new_geovctrs_wkt}
\alias{is_geovctrs_wkt}
\alias{validate_geovctrs_wkt}
\alias{vec_cast.geovctrs_wkt}
\alias{vec_ptype2.geovctrs_wkt}
\title{S3 details for geovctrs_wkt}
\usage{
new_geovctrs_wkt(
  x = character(),
  trim = TRUE,
  precision = 16L,
  dimensions = 3L
)

is_geovctrs_wkt(x)

validate_geovctrs_wkt(x)

\method{vec_cast}{geovctrs_wkt}(x, to, ...)

\method{vec_ptype2}{geovctrs_wkt}(x, y, ...)
}
\arguments{
\item{x}{A character vector containing well-known text}

\item{trim}{Trim unnecessary zeroes in the output?}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{dimensions}{The maximum number of dimensions to consider
when generating text.}

\item{...}{Unused}

\item{y, to}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geovctrs_wkt
}
\examples{
wkt <- geo_wkt("POINT (30 10)")
is_geovctrs_wkt(wkt)

# use new_geovctrs_wkt() to skip parse validation if you know
# your text is valid WKT
new_geovctrs_wkt("POINT (30 10)")

}
