% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-xy.R, R/geo-xyz.R
\name{new_geovctrs_xy}
\alias{new_geovctrs_xy}
\alias{is_geovctrs_xy}
\alias{as_geo_xy.matrix}
\alias{as.matrix.geovctrs_xy}
\alias{as_geo_xy}
\alias{as_geo_xy.default}
\alias{vec_cast.geovctrs_xy}
\alias{vec_ptype2.geovctrs_xy}
\alias{new_geovctrs_xyz}
\alias{is_geovctrs_xyz}
\alias{as_geo_xyz.matrix}
\alias{as_geo_xyz}
\alias{as_geo_xyz.default}
\alias{vec_cast.geovctrs_xyz}
\alias{vec_ptype2.geovctrs_xyz}
\title{S3 details for geovctrs_xy}
\usage{
new_geovctrs_xy(x = list(x = double(), y = double()))

is_geovctrs_xy(x)

\method{as_geo_xy}{matrix}(x, ...)

\method{as.matrix}{geovctrs_xy}(x, ...)

as_geo_xy(x, ...)

\method{as_geo_xy}{default}(x, ...)

\method{vec_cast}{geovctrs_xy}(x, to, ...)

\method{vec_ptype2}{geovctrs_xy}(x, y, ...)

new_geovctrs_xyz(x = list(x = double(), y = double(), z = double()))

is_geovctrs_xyz(x)

\method{as_geo_xyz}{matrix}(x, ...)

as_geo_xyz(x, ...)

\method{as_geo_xyz}{default}(x, ...)

\method{vec_cast}{geovctrs_xyz}(x, to, ...)

\method{vec_ptype2}{geovctrs_xyz}(x, y, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_xy]{geo_xy()}}}

\item{...}{Unused}

\item{to, y}{Arguments to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 details for geovctrs_xy
}
