% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-plot.R
\name{plot.geovctrs_xy}
\alias{plot.geovctrs_xy}
\alias{plot.geovctrs_segment}
\alias{plot.geovctrs_rect}
\title{Plot geometry vectors}
\usage{
\method{plot}{geovctrs_xy}(
  x,
  ...,
  add = FALSE,
  asp = 1,
  bbox = geo_bbox(x, finite = TRUE),
  xlab = "",
  ylab = ""
)

\method{plot}{geovctrs_segment}(
  x,
  ...,
  add = FALSE,
  asp = 1,
  bbox = geo_bbox(x, finite = TRUE),
  xlab = "",
  ylab = ""
)

\method{plot}{geovctrs_rect}(
  x,
  ...,
  add = FALSE,
  asp = 1,
  bbox = geo_bbox(x, finite = TRUE),
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Passed to plotting functions for features: \code{\link[graphics:points]{graphics::points()}}
for xy vectors, \code{\link[graphics:segments]{graphics::segments()}} for segment vectors, and
\code{\link[graphics:rect]{graphics::rect()}} for rectangles.}

\item{add}{Should the object be added the current plot?}

\item{asp, xlab, ylab}{Passed to \code{\link[graphics:plot.default]{graphics::plot()}}}

\item{bbox}{The limits of the plot. Defaults to \code{geo_bbox(x, finite = TRUE)}.}
}
\value{
\code{x}, invisibly.
}
\description{
These functions are basic plot implementations for the geovctrs types.
Similar \code{\link[=plot]{plot()}} methods are available for \link[wkutils:wkt_plot]{well-known types}
in the wkutils package.
}
\examples{
plot(geo_xy(1:5, 1:5))
plot(geo_segment(0, 0, 1, 1))
plot(geo_rect(0, 1, 2, 3), col = "grey90")

}
