% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate-unite.R
\name{unite_xy}
\alias{unite_xy}
\alias{separate_xy}
\title{Unite and separate components of geometries}
\usage{
unite_xy(data, col, x, y, remove = TRUE)

separate_xy(data, col, into = c("x", "y"), remove = TRUE)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble::tibble()}} or data frame}

\item{col}{A column name for the \code{\link[=geo_xy]{geo_xy()}} object}

\item{x, y}{Column specifications for x and y coordinate
columns, respectively. Use \code{\link[dplyr:select]{dplyr::select()}} syntax.}

\item{remove}{Use \code{remove = FALSE} to keep \code{x} and \code{y} as columns
in the output.}
}
\value{
\code{data}, with a new column \code{col}
}
\description{
These functions work like \code{\link[tidyr:separate]{tidyr::separate()}} and \code{\link[tidyr:unite]{tidyr::unite()}},
combining information from multiple columns into single column.
}
\examples{
tbl <- tibble(x = 1, y = 2)
united <- unite_xy(tbl, "xy", x, y)
united

separate_xy(united, xy, into = c("x", "y"))

}
