% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-coord.R
\name{vec_ptype2.geo_coord}
\alias{vec_ptype2.geo_coord}
\alias{vec_cast.geo_coord}
\title{Vctrs details for coordinate vectors}
\usage{
\method{vec_ptype2}{geo_coord}(x, y, ...)

\method{vec_cast}{geo_coord}(x, to, ...)
}
\arguments{
\item{x, y, to, ...}{See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
Coordinate vectors are an internal data structure
that will probably be phased out in future versions.
}
