% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-bbox.R
\name{geo_bbox}
\alias{geo_bbox}
\alias{geo_xlim}
\alias{geo_ylim}
\alias{geo_envelope}
\title{Bounding boxes and envelopes}
\usage{
geo_bbox(x, ..., na.rm = FALSE, finite = FALSE)

geo_xlim(x, ..., na.rm = FALSE, finite = FALSE)

geo_ylim(x, ..., na.rm = FALSE, finite = FALSE)

geo_envelope(x, ..., na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Unused}

\item{na.rm}{Should NAs be removed?}

\item{finite}{Should only finite values be considered? \code{TRUE}
implies \code{na.rm = TRUE}.}
}
\value{
A \code{\link[=geo_rect]{geo_rect()}} of length 1.
}
\description{
Whereas a \code{\link[=geo_bbox]{geo_bbox()}} of a \link[=is_geovctr]{geovctr} is always
of length 1, a \code{\link[=geo_envelope]{geo_envelope()}} shares the same length as
the vector. Both return a \code{\link[=geo_rect]{geo_rect()}}. Empty vectors
(and empty geometries) return \code{geo_rect(Inf, Inf, -Inf, -Inf)},
and \code{NA/NaN} values are removed if \code{na.rm = TRUE} (which might
mean more \code{Inf} values than you expected).
}
\examples{
geo_bbox(geo_wkt(c("POINT (30 10)", "POINT EMPTY")))
geo_envelope(geo_wkt(c("POINT (30 10)", "POINT EMPTY")))

}
