% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-collection.R
\name{geo_collection}
\alias{geo_collection}
\title{Collections of coordinate vectors}
\usage{
geo_collection(feature = list(), srid = 0)
}
\arguments{
\item{feature}{A list of one or more features.}

\item{srid}{A spatial reference identifier}
}
\value{
A \code{\link[=new_geo_collection]{new_geo_collection()}}
}
\description{
Collections of coordinate vectors
}
