% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-multipoint.R
\name{geo_multipoint}
\alias{geo_multipoint}
\title{Create a multipoint}
\usage{
geo_multipoint(feature, srid = NA)
}
\arguments{
\item{feature}{A \code{\link[=geo_collection]{geo_collection()}} of \code{\link[=geo_point]{geo_point()}}s.}

\item{srid}{A spatial reference identifier.}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a multipoint
}
\examples{
geo_multipoint(
  geo_point(geo_xy(10, 30))
)

}
