% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-multipoint.R
\name{geo_multipoint}
\alias{geo_multipoint}
\title{Create a multipoint}
\usage{
geo_multipoint(xy, srid = NA)
}
\arguments{
\item{xy}{A \code{\link[=geo_xy]{geo_xy()}}}

\item{srid}{A spatial reference identifier.}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a multipoint
}
\examples{
geo_multipoint(geo_xy(1:5, 2:6))

}
