% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-multipolygon.R
\name{geo_multipolygon}
\alias{geo_multipolygon}
\title{Create a multipolygon}
\usage{
geo_multipolygon(feature, srid = NA)
}
\arguments{
\item{feature}{A \code{\link[=geo_collection]{geo_collection()}} of \code{\link[=geo_polygon]{geo_polygon()}}s.}

\item{srid}{A spatial reference identifier.}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a multipolygon
}
\examples{
geo_multipolygon(
  geo_polygon(geo_xy(c(0, 10, 0), c(0, 0, 10)))
)

}
