% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-point.R
\name{geo_point}
\alias{geo_point}
\title{Create a point}
\usage{
geo_point(xy, srid = 0)
}
\arguments{
\item{xy}{A \code{\link[=geo_xy]{geo_xy()}}}

\item{srid}{A spatial reference identifier.}
}
\value{
A \code{\link[=geo_collection]{geo_collection()}} of length 1.
}
\description{
Create a point
}
\examples{
geo_point(geo_xy(0, 0))

}
