% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-segment.R
\name{geo_segment}
\alias{geo_segment}
\title{Line segments}
\usage{
geo_segment(start = geo_xy(), end = geo_xy(), srid = 0)
}
\arguments{
\item{start, end}{\code{\link[=geo_xy]{geo_xy()}}s for the start and end
of the segment, respectively.}

\item{srid}{A spatial reference identifier, coerced to
an integer by \code{\link[=as_geo_srid]{as_geo_srid()}}. These identifiers can
and should be managed outside of geovctrs except for
0, which is interpreted as "not set".}
}
\value{
A \code{\link[=new_geo_segment]{new_geo_segment()}}
}
\description{
The \code{\link[=geo_segment]{geo_segment()}} type is useful as an efficient representation of
line segments stored using column vectors.
Note that \code{geo_segment(geo_xy(NA, NA), geo_xy(NA, NA))} is considered
an "empty" linestring, whereas \code{geo_segment(geo_xy(NA, NA), geo_xy(NA, NA), srid = NA)}
is "missing" (see \code{\link[=geo_is_missing]{geo_is_missing()}} and \code{\link[=geo_is_empty]{geo_is_empty()}}).
}
\examples{
geo_plot(geo_segment(geo_xy(0, 0), geo_xy(10, -10:10)))

}
