% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-srid.R
\name{geo_srid}
\alias{geo_srid}
\alias{geo_srid.default}
\alias{geo_set_srid}
\alias{geo_set_srid.default}
\alias{as_geo_srid}
\title{Extract a spatial reference identifier}
\usage{
geo_srid(x)

\method{geo_srid}{default}(x)

geo_set_srid(x, srid)

\method{geo_set_srid}{default}(x, srid)

as_geo_srid(x)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{srid}{A spatial reference identifier, coerced to
an integer by \code{\link[=as_geo_srid]{as_geo_srid()}}. These identifiers can
and should be managed outside of geovctrs except for
0, which is interpreted as "not set".}
}
\value{
An integer vector (one SRID per feature).
}
\description{
The geovctrs package works with spatial reference identifiers
instead of actual CRS objects so that the heavy lifting can
be implemented in other packages. However, most geometry
formats can store an integer spatial reference identifier (SRID)
with each feature to propagate this information through
calculations.
}
\examples{
# two points with an SRID
geometries <- wkt(
  c(
    "SRID=26920;POINT (259473 4876249)",
    "SRID=4326;POINT (-66 44)"
  )
)

geo_srid(geometries)
geo_srid(geo_set_srid(geometries, NA))

# SRIDs are propogated through conversions,
# or discarded with a warning
geo_srid(as_wkb(geometries))
geo_srid(as_geo_xy(geometries))

}
