% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-summary.R
\name{geo_summary}
\alias{geo_summary}
\alias{geo_geometry_type}
\title{Feature-level geometry summaries}
\usage{
geo_summary(x)

geo_geometry_type(x)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{is_empty}, \code{geometry_type},
\code{n_geometries}, \code{n_coordinates}, \code{srid}, \code{coordinate_dimensions},
\code{has_z}, \code{first_coordinate}, \code{problems}, and \code{is_missing}.
}
\description{
Feature-level geometry summaries
}
\examples{
geo_summary(geo_nc)

geo_summary(
  geo_wkt(
    c(
      "POINT (30 10)",
      "LINESTRING (0 0, 10 11)",
      "POLYGON EMPTY"
    )
  )
)

}
