% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-xy.R, R/geo-xyz.R
\name{geo_xy}
\alias{geo_xy}
\alias{geo_xyz}
\title{Create a coordinate vector}
\usage{
geo_xy(x = double(), y = double())

geo_xyz(x = double(), y = double(), z = double())
}
\arguments{
\item{x, y, z}{x, y, and z coordinate vectors}
}
\value{
A \code{\link[=new_geovctrs_xy]{new_geovctrs_xy()}}
}
\description{
The \code{\link[=geo_xy]{geo_xy()}} type is useful as an efficient representation of points
stored using column vectors. Note that \code{geo_xy(NA, NA)} is considered
an "empty" point rather than a "missing" point (see \code{\link[=geo_is_missing]{geo_is_missing()}}
and \code{\link[=geo_is_empty]{geo_is_empty()}}).
}
\examples{
geo_xy(0:5, 1:6)
plot(geo_xy(0:5, 1:6))

geo_xyz(0:5, 1:6, 3)

}
