% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-collection.R
\name{new_geo_collection}
\alias{new_geo_collection}
\alias{validate_geo_collection}
\alias{is_geo_collection}
\alias{as_geo_collection}
\alias{as_geo_collection.default}
\alias{vec_cast.geovctrs_collection}
\alias{vec_ptype2.geovctrs_collection}
\title{S3 Details for coordinate vector collections}
\usage{
new_geo_collection(x = list(feature = list(), srid = integer()))

validate_geo_collection(x)

is_geo_collection(x)

as_geo_collection(x, ...)

\method{as_geo_collection}{default}(x, ...)

\method{vec_cast}{geovctrs_collection}(x, to, ...)

\method{vec_ptype2}{geovctrs_collection}(x, y, ...)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_collection]{geo_collection()}}}

\item{...}{Unused}

\item{to}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}

\item{y}{A prototype to cast to. See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}}
}
\description{
S3 Details for coordinate vector collections
}
