% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-collection.R
\name{new_geo_collection}
\alias{new_geo_collection}
\alias{validate_geo_collection}
\alias{is_geo_collection}
\title{S3 Details for coordinate vector collections}
\usage{
new_geo_collection(x = list(feature = list(), srid = integer()))

validate_geo_collection(x)

is_geo_collection(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=geo_collection]{geo_collection()}}}
}
\description{
S3 Details for coordinate vector collections
}
