% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-plot.R
\name{plot.geovctrs_xy}
\alias{plot.geovctrs_xy}
\alias{plot.geovctrs_segment}
\alias{plot.geovctrs_rect}
\title{Plot geometry vectors}
\usage{
\method{plot}{geovctrs_xy}(
  x,
  ...,
  add = FALSE,
  asp = 1,
  bbox = geo_bbox(x, finite = TRUE),
  xlab = "",
  ylab = ""
)

\method{plot}{geovctrs_segment}(
  x,
  ...,
  add = FALSE,
  asp = 1,
  bbox = geo_bbox(x, finite = TRUE),
  xlab = "",
  ylab = ""
)

\method{plot}{geovctrs_rect}(
  x,
  ...,
  add = FALSE,
  asp = 1,
  bbox = geo_bbox(x, finite = TRUE),
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{x}{A geometry-like object, or one that can be
coerced to a geometry-like object using \code{\link[=as_geovctr]{as_geovctr()}}.}

\item{...}{Passed to plotting functions for features: \code{\link[graphics:points]{graphics::points()}}
for xy vectors, \code{\link[graphics:segments]{graphics::segments()}} for segment vectors, and
\code{\link[graphics:rect]{graphics::rect()}} for rectangles.}

\item{add}{Should the object be added the current plot?}

\item{asp, xlab, ylab}{Passed to \code{\link[graphics:plot.default]{graphics::plot()}}}

\item{bbox}{The limits of the plot. Defaults to \code{geo_bbox(x, finite = TRUE)}.}
}
\value{
\code{x}, invisibly.
}
\description{
These functions are basic plot implementations for the geovctrs types.
Similar \code{\link[=plot]{plot()}} methods are available for \link[wkutils:plot.wk_wkt]{well-known types}
in the wkutils package.
}
\examples{
# plot functions for wk types are provided
# in the wk package:
plot(wkt("POINT (10 40)"))
plot(wkt("LINESTRING (30 10, 10 30, 40 40)"))
plot(wkt("MULTIPOINT ((10 40), (40 30))"))
plot(wkt("MULTILINESTRING ((10 10, 20 20, 10 40), (40 40, 30 30, 40 20, 30 10))"))
plot(wkt("POLYGON ((30 10, 10 30, 40 40, 30 10))"))
plot(
  wkt(
    "MULTIPOLYGON (((40 40, 20 45, 45 30, 40 40)),
      ((20 35, 10 30, 10 10, 30 5, 45 20, 20 35), (30 20, 20 15, 20 25, 30 20)))"
  ),
  col = "grey90"
)

# plot functions with the same arguments are also available:
plot(geo_xy(1:5, 1:5))
plot(geo_segment(0, 0, 1, 1))
plot(geo_rect(0, 1, 2, 3), col = "grey90")

}
