% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geovctrs-package.R, R/wk.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tibble}
\alias{as_wkt}
\alias{wkt}
\alias{new_wk_wkt}
\alias{parse_wkt}
\alias{as_wkb}
\alias{wkb}
\alias{new_wk_wkb}
\alias{parse_wkb}
\alias{as_wksxp}
\alias{wksxp}
\alias{new_wk_wksxp}
\alias{parse_wksxp}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{wk}{\code{\link[wk]{as_wkb}}, \code{\link[wk]{as_wksxp}}, \code{\link[wk]{as_wkt}}, \code{\link[wk]{new_wk_wkb}}, \code{\link[wk]{new_wk_wksxp}}, \code{\link[wk]{new_wk_wkt}}, \code{\link[wk]{parse_wkb}}, \code{\link[wk]{parse_wksxp}}, \code{\link[wk]{parse_wkt}}, \code{\link[wk]{wkb}}, \code{\link[wk]{wksxp}}, \code{\link[wk]{wkt}}}
}}

